Split provider orchestration by backend
